Add support for TTLs to token roles
We recently became interested in using the Token auth backend roles.

However, our use-case has been stymied, because the Token auth method
lacks support for per-role configuration of `token_ttl` and
`token_max_ttl`, unlike other major auth backends.

There doesn't seem to be an obvious reason for this, and it looks fairly
easy to plumb through the parameters to where they need to be used.