Remove browserstack dep and tests (#18432)
This PR aims to remove the browserstack test from the CI pipeline. This was prompted by known failures, relating to legacy support. Since we no longer support IE11 for Vault back to version 1.9.x, we can drop the browserstack dependency and related tests.

This was originally done in #15997 for 1.12.x+, so just bringing it back to 1.11.x to drop the noisy failures.