Add support for `prefix` in password policies
It is an industry standard practice to generate secrets that include an
identifying fixed prefix, to make it easier to identify secrets that are
mistakenly leaked into logging and version control systems.

This article on the GitHub blog describes the practice, including links
to Slack and Stripe documentation showing those vendors also observe
this practice: https://github.blog/2021-04-05-behind-githubs-new-authentication-token-formats/

Another widely known example is that Amazon AWS access key IDs start
with `AKIA` (or `ASIA` for short-lived credentials).

The Vault Active Directory secrets engine accidentally implements
the practice already, in its deprecated password generation routine, by
starting all passwords for `?@09AZ` - "accidentally", because the stated
reason is to work around password complexity rules on the server side,
but it also accomplishes this goal too.

However the newer "Password Policies" feature in Vault is incapable of
incorporating fixed prefixes on passwords.

This poses a problem for people who would like to migrate away from
deprecated features - since the warnings generated by
`terraform-provider-vault` are irritatingly unsuppressable - but are
unwilling to sacrifice this security feature.

Therefore, this PR adds in capability to specify a fixed prefix to the
new "Password Policies" generation system.