Add events sending routed from plugins
This isn't perfect, for sure, as handling of namespaces is a bit
rudimentary, but it's solidifying and becoming a useful base to work off.

This routes events sent from auth and secrets plugins to the main
`EventBus` in the Vault Core. Events sent from plugins are automatically
tagged with the namespace and plugin information associated with them.