Make it possible to override the builtin UI
This change adds a new configuration option that makes it possible to
override the builtin web UI. This makes it possible to fix bugs, or add
support for new features or internal secret engines before they are
officially released.

Overriding the builtin UI is already easy for users of Vault
Community Edition but there is no way to do so for users of the
Enterprise Edition.

Consul already has a [similar option](https://developer.hashicorp.com/consul/docs/agent/config/config-files#ui_config_dir)
so it should not be out of place for Vault to have the same possibility.

In this change I marked the `ui` option as deprecated, it is now an
alias of `ui_config.enabled`. Having a global `ui_config` block will
make it easy to add more parameter later if needed, like the equivalent
of Consul's `content_path` which [has been requested by some users](https://github.com/hashicorp/vault/issues/9221).