events: Keep a small in-memory replay buffer
The most recent 32 events are kept in a deque buffer,
and replayed to new subscribers.

We use a RWLock to control access. Writing a new event will block
all other access, since it needs to modify some of the internal deque
state. However, since we carefully control the size of the deque,
the lock time should be minimal.

When a new subscriber is added, the replay buffer is copied (as a slice
of pointers to the events) before the events are sent to the new receiver
channel. The copy is to minimize the time that the read lock is held
even if the receiving channel is slow to process the messages.