[QT-525] enos: use spot instances for Vault targets
The previous strategy for provisioning infrastructure targets was to use
the cheapest instances that could reliably perform as Vault cluster
nodes. With this change we introduce a new model of bidding for spot
fleet instances with the goal of costs savings and often more powerful
instances. The spot fleet instance bidding has only been implemented for
Vault clusters. Updating our Consul backend bidding will be handled in
another PR.

* Create a new `vault_cluster` module that handles installation,
  configuration, initializing, and unsealing Vault clusters.
* Create a `target_ec2_instances` module that can provision a group of
  instances on-demand.
* Create a `target_ec2_spot_fleet` module that can bid on a fleet of
  spot instances.
* Extend every Enos scenario to utilize the spot fleet target acquisition
  strategy and the `vault_cluster` module.
* Update our Enos CI modules to handle both the `aws-nuke` permissions
  and also the privileges to provision spot fleets.

Signed-off-by: Ryan Cragun <me@ryan.ec>