api: remove direct dependency on hclog
In tests, define a local implementation of a null logger satisfying
the retryablehttp.LeveledLogger interface as a replacement of a
reference to hclog.NewNullLogger(). This allows to drop the single
direct reference to hclog and so to reduce dependencies for a leaner impact on
downstream modules.
Related to #14017 (too many dependencies).