Always enable event system for PKI
We are considering using the event system for internal purposes
even if it is not enabled in general.

This adds the option to selectively enable the event system for
particular secrets engines (currently set only to `pki`).

This should still have minimal impact if not globally enabled,
as the event bus will not even be populated in secrets engines not
in the enabled map, and additionally events will be rejected by the
event bus when attempting to send.

I tried a few ways to accomplish this, such as specifically allowing
only certain event types, but this is what I ultimately found to be the
cleanest solution with the fewest surprises.