Add ability to drain ACME validation queue
When disabling ACME, drain the validation queue so that no further work
items are processed. Later, when ACME is re-enabled, read queue entries
from storage so we re-attempt anything that might've been drained.

Because the queue is storage-first, these operations at worst re-do some
validation that was already performed (if the cluster went down before
removing a finished validation).

Signed-off-by: Alexander Scheel <alex.scheel@hashicorp.com>